requirements.txt
streamlit
python-dotenv
google-generativeai
requests
paho-mqtt